56. fix synthetic data generation and reorganize test structure
- Rename testing module to testdata for clarity
- Rename tests/ to test/ directory
- Add matplotlib dependency for annotation image generation
- Generate uint8 images with improved noise levels
- Add ground truth annotation images with bounding boxes
- Update pyproject.toml and pytest configuration

+++ src/python/test/test_synthetic.py
@@ -44,7 +44,7 @@ def test_mini_image_loadable(mini_dataset: Path, mini_ground_truth: dict):
     image = tifffile.imread(tiff_path)
 
     assert image.shape == tuple(shape)
-    assert image.dtype == np.uint16
+    assert image.dtype == np.uint8
 
 
 def test_codebook_matches_ground_truth(mini_codebook: dict, mini_ground_truth: dict):
